+++ src/Symfony/Component/DependencyInjection/Compiler/PassConfig.php
@@ -43,12 +43,13 @@ class PassConfig
             100 => array(
                 new ResolveClassPass(),
                 new ResolveInstanceofConditionalsPass(),
+                new RegisterEnvVarProcessorsPass(),
             ),
         );
 
         $this->optimizationPasses = array(array(
             new ExtensionCompilerPass(),
-            new ResolveDefinitionTemplatesPass(),
+            new ResolveChildDefinitionsPass(),
             new ServiceLocatorTagPass(),
             new DecoratorServicePass(),
             new ResolveParameterPlaceHoldersPass(false),