+++ src/Symfony/Component/DependencyInjection/Tests/Fixtures/php/services_base64_env.php
@@ -9,14 +9,14 @@ use Symfony\Component\DependencyInjection\Exception\RuntimeException;
 use Symfony\Component\DependencyInjection\ParameterBag\FrozenParameterBag;
 
 /**
- * Symfony_DI_PhpDumper_Test_EnvParameters.
+ * Symfony_DI_PhpDumper_Test_Base64Parameters.
  *
  * This class has been auto-generated
  * by the Symfony Dependency Injection Component.
  *
  * @final since Symfony 3.3
  */
-class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
+class Symfony_DI_PhpDumper_Test_Base64Parameters extends Container
 {
     private $parameters;
     private $targetDirs = array();
@@ -27,16 +27,9 @@ class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
      */
     public function __construct()
     {
-        $dir = __DIR__;
-        for ($i = 1; $i <= 5; ++$i) {
-            $this->targetDirs[$i] = $dir = dirname($dir);
-        }
         $this->parameters = $this->getDefaultParameters();
 
         $this->services = $this->privates = array();
-        $this->methodMap = array(
-            'test' => 'getTestService',
-        );
 
         $this->aliases = array();
     }
@@ -64,18 +57,6 @@ class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
     public function isCompiled()
     {
         return true;
-    }
-
-    /**
-     * Gets the public 'test' shared service.
-     *
-     * @return object A %env(FOO)% instance
-     */
-    protected function getTestService()
-    {
-        $class = $this->getEnv('FOO');
-
-        return $this->services['test'] = new $class($this->getEnv('Bar'), 'foo'.$this->getEnv('string:FOO').'baz', $this->getEnv('int:Baz'));
     }
 
     /**
@@ -130,11 +111,7 @@ class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
     }
 
     private $loadedDynamicParameters = array(
-        'bar' => false,
-        'baz' => false,
-        'json' => false,
-        'db_dsn' => false,
-        'env(json_file)' => false,
+        'hello' => false,
     );
     private $dynamicParameters = array();
 
@@ -150,11 +127,7 @@ class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
     private function getDynamicParameter($name)
     {
         switch ($name) {
-            case 'bar': $value = $this->getEnv('FOO'); break;
-            case 'baz': $value = $this->getEnv('int:Baz'); break;
-            case 'json': $value = $this->getEnv('json:file:json_file'); break;
-            case 'db_dsn': $value = $this->getEnv('resolve:DB'); break;
-            case 'env(json_file)': $value = ($this->targetDirs[1].'/array.json'); break;
+            case 'hello': $value = $this->getEnv('base64:foo'); break;
             default: throw new InvalidArgumentException(sprintf('The dynamic parameter "%s" must be defined.', $name));
         }
         $this->loadedDynamicParameters[$name] = true;
@@ -170,9 +143,7 @@ class Symfony_DI_PhpDumper_Test_EnvParameters extends Container
     protected function getDefaultParameters()
     {
         return array(
-            'project_dir' => '/foo/bar',
-            'env(FOO)' => 'foo',
-            'env(DB)' => 'sqlite://%project_dir%/var/data.db',
+            'env(foo)' => 'd29ybGQ=',
         );
     }
 }